Properties after thesis
Property 6 in thesis, which was proved only mathematically is the thesis, is proved with Coq as well. It is called NegativeWeights in the file.
NegativeCurrent and AllNegative are two lemmas, which are proved to help proving NegativeWeights.
Currently working on four lemmas, which are
NegativeLoop00Next00
NegativeLoop00Next01
NegativeLoop00Next10
NegativeLoop00Next11
To be able to prove properties 10 and 11 in the thesis.